apps/worker/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
prometheus-client==0.19.0
flower==2.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1
httpx==0.25.2
structlog==23.2.0
tenacity==8.2.3
circuitbreaker==2.0.0